refactor(api): move file replace to public API as replaceFile
Move the in-place file replacement out of the plugin files API and onto the
public UploaderPublicApi as `replaceFile` (remove + add at the same position,
preserving the source's context). The plugin `files` API now exposes only the
plugin-specific `update`; plugins reach add/remove/replaceFile through the
public `uploaderApi` they receive in setup. Update the e2e test to drive the
public `api.replaceFile`.
