+++ src/abstract/managers/plugin/buildPluginApi.ts
@@ -3,12 +3,8 @@ import type { SharedState } from '../../../lit/SharedState';
 import type { SharedInstancesBag } from '../../../lit/shared-instances';
 import type { Uid } from '../../../lit/Uid';
 import type { ConfigType } from '../../../types';
-import type { UploadcareFile } from '../../../types/exported';
-import { UploadSource } from '../../../utils/UploadSource';
 import type { CustomConfig } from '../../customConfigOptions';
 import { sharedConfigKey } from '../../sharedConfigKey';
-import type { ApiAddFileCommonOptions } from '../../UploaderPublicApi';
-import { uploadcareFileToEntryData } from '../../uploadEntrySchema';
 import type { PluginRegistry } from './PluginRegistry';
 import type {
   PluginActivityApi,
@@ -86,36 +82,6 @@ export function buildPluginApi(
       if (changes.cdnUrl !== undefined) entry.setValue('cdnUrl', changes.cdnUrl);
       if (changes.cdnUrlModifiers !== undefined) entry.setValue('cdnUrlModifiers', changes.cdnUrlModifiers);
       if (changes.mimeType !== undefined) entry.setValue('mimeType', changes.mimeType);
-    },
-    replace: (internalId: string, file: UploadcareFile, { silent, fileName, source }: ApiAddFileCommonOptions = {}) => {
-      const oldId = internalId as Uid;
-      const collection = sharedInstancesBag.uploadCollection;
-      const index = collection.items().indexOf(oldId);
-      const oldEntry = collection.read(oldId);
-      if (index === -1 || !oldEntry) {
-        throw new Error(`File with internalId ${internalId} not found`);
-      }
-      // Carry over the original entry's context so the replacement keeps its
-      // identity within the session; everything file-specific comes from `file`.
-      const preserved = {
-        source: oldEntry.getValue('source'),
-        metadata: oldEntry.getValue('metadata'),
-        fullPath: oldEntry.getValue('fullPath'),
-        silent: oldEntry.getValue('silent'),
-      };
-      collection.remove(oldId);
-      const newId = collection.add(
-        {
-          ...uploadcareFileToEntryData(file),
-          fileName: fileName ?? file.originalFilename ?? null,
-          silent: silent ?? preserved.silent,
-          source: source ?? preserved.source ?? UploadSource.API,
-          metadata: preserved.metadata,
-          fullPath: preserved.fullPath,
-        },
-        { index },
-      );
-      return sharedInstancesBag.api.getOutputItem<'success'>(newId);
     },
   };
